Decided to add Hulu to the default GoogleTV conf.
To really use it effectively one needs to either have a mouse
or have (on Linux) lircmd configured.  I have done that for my
machines and it works fairly well.  In the RC package we tried
to have our own code control the mouse (which would be X-platform)
but it just doesn't perform well as it will just get "stuck".