Fix a few bugs with the new VideoScreen.
The filtering of TV show episodes was off because it wasn't
checking if the compared Video was actually a TV episode.  So
a movie of the same name would be included.

Also a display bug when switching TV seasons if the user has changed
the episode selection to an index not existing in the new list.
Now set it to the first episode.