Sometimes it's handy not to send keyboard commands.
The google tv "sites" are supposed to be arrow key driven
but sometimes one wants to add a site that is not been
redesigned for google tv.  Like Hulu for example.  It's
nice to launch Chrome to display it but you need the
mouse to control it.  Matter of fact sending arrow keys
kinds of messes it up actually.

So in googletv.properties one can add:

",false" to one of the properties to turn off sending keyboard
commands.  It defaults to true so you have to explicitly turn
it off if that is what you want.

Here in the lab we run Lircmd to control the mouse nicely and
a button on the remote configured to turn mouse control on
and off.