Make the popup menus nicer - don't use Java Dialogs.
At this point we have all but the RecordingRule Panel
converted to more rich looking UI panels.  The new popups
are from the revamped ButtonPanel class - It now fades
in on starting and camn display an image (currently using
the logo).  Also it's background is a rounded rectangle
using the theme panel color and alpha.

Looks good just being a non rectangle.