Supply simple banner backrounds.
The google tv and on demand screens now have a simple banner
window and we made some basic image backgrounds.

Also zapped warping the mouse on startup - not needed and a hack
because some popups didn't redefine the mouse so it was seen.

Now because mouse usage is optional using chrome, it's best just to
not futz with by warping.