Set the RecordingRulePanel to theme colors.
I was going to rewrite it to be more of a home grown
UI panel and pop it up like the ButtonPanel but it is
quite a complicated UI.  I didn't like any of the
solutions that I can up with doing.

So I kept the Dialog popup but at least it uses the
proper colors, and I follow focus to set the highlight.
This will need to be good enough for now.  Perhaps I
could add the logo image to make it look a bit better.

Of course no fade in since we don't have top-level
transparency in our Windows.