Many updates - indexed files, worker cleanup.
Now a Recorder can suggest a Worker (by name) that could do
a nice job on converting the transport stream to some sort
of indexed file (mkv, mp4 etc).  These files seek much nicer
using mplayer2.  It increases the need to disk space as we
keep both versions of the file.  We don't delete because as of
now we don't keep track of clients and we don't want to delete
a file from underneath them.

Cleaned up the worker code as much code was duplicated.  Now making
a new worker takes a lot less code.

Started work on our own commercial detection code.  We have no
illusions that we can be better than comskip but comskip does
pretty poor on our HD-PVR mp4 transport streams.  So we thought
we could improve functionality there looking for the "rating symbol"
that is displayed when the show starts again.  Still need to make
"workers" that can be chosen when making a RecordingRule.