Looks like the HD-PVR recording is now very robust.
I found when the machine is very IO intense the
reading from the /dev/videoN device for the HD-PVR
can "lose" it's connection - meaning the recording
will stop yet we will still be "blocked" on our
read call.

What used to happen is that this block will continue
for the rest of the recording.  What we do now is have
a timer check to ensure the recording is happening
and if necessary interrupt the block and restart.  We
do lose some recording time (up to 10 seconds?) but this
is better that losing the whole thing.  After much testing
we have had zero failures on recordings from the HS-PVR.

We declare it now to be very reliable.