We reworked the detect rating symbol code.
I decided that just looking for a generic box was not quite
good enough because too many false positives.  So we look for
actual images.  We have a directory of images to use that always
can be added to to cover more shows.  When the appropriate image
is included the code has been perfect.  No misses and no false
positives.  The downside will be keeping up with the images
because it seems like they keep changing them on purpose just
to foil detection.