We want to sort videos/recordings properly.
Up until now we just sorted by the title as a String.  This
puts everything with "The" together which really isn't
correct.  So we made a Util method to handle title Strings
so "The" is eliminated and that output string is used to
sort.  At this point it's just "The" being handled but since
we isolated this code to a utility method we can add code there
and in all the screens that sort will use the method and pick
up enhancements that way.