The scheduling lists are now sorted properly.
We don't want all the shows to be sorted by "The", in
other words sorting is now handled properly for the
scheduling screens.  Also the toString() method of the
ShowAiring object made a String that sorted poorly.
For example going from September to October data, the
the October data would be sorted first since "10" does
come before "9".  Now it won't because it will be "09"
and "10".  Plus it will begin with the Year so things
will alos work from December to January.