Several updates to make things a bit smoother.
The workers run after a recording complete now run with
ionice.  I had tested with nice but didn't see that it made any
difference.  I didn't know aboutr ionice until I googled for it.
The hope is that IO resources will not be swamped now and the
video glitches using the HD-PVR will be fixed.  So far so good.

I added two more remote commands to the Roku on demand.  The main
one being the "replay" function.

The NMS now has a new property called GroupName.  This
allows clients to be able to filter the NMS instances.  I
needed this so we can run production and test servers.  This
should make testing new stuff without affecting the other TVs
in the house.