A fairly large set of changes.
Foremost is support for the Roku as a client.  To
which the default conf files were changed to support
the Roku setup.  Just easier to have things default
to a server meant to be used with the Roku.

When the user has not set the IP of the host via
/etc/hosts, then our discovery network code didn't
work right.  Up until now I have just set it in the
host file.  On Ubuntu this is not the default so
a better way is needed.  Now we will also look in
/etc/network/interfaces for static address for eth0.
This is not perfect but better.

The SD ProgramData now will only connect if the XTVD.xml
file has been initialized with user data.  That file is
now updated when the user updates the Configuration for
the SD ProgramData service.