We have added an "auto art" service.
This service can get artwork for recordings and videos
without user interaction.  Hopefully will be a time
saver and not a waster.  It is not perfect and the user
may have to undo some images from time to time.  They
would use the desktop app to fix things.

Also added some code for the Roku client.  The roku caches
image URLs so when they update it won't re-fetch them.  So
now we don't put artwork images in the XML if they don't
actually exist.