Many updates to add some polish to our system.
Created a new app called "ez" which is meant to
be a simpler to use app to configure a jflicks
server.

Updated the config files for the desktop apps.

The Bif code didn't work in our turnkey linux
distro because of missing fonts.  So I changed the
code to use images instead that are included in
the bundle.  This way we can avoid font issues
which makes sense since we are a server.

Made the schedules direct get new data on it's next
check after the user has changed the user/password.