We are now fully capable to run without NFS.
We wanted to be able to not need NFS for the system to
work.  All media and data is available over the network.
If users want to use NFS it will work and be the first
choice.  Actually the first choice is using the file directly,
not necessary via NFS.  I mean a stand alone system would
just be local files.

Added some notions to the Worker classes so we can do things
like group commercial flaggers as we want to select just one
of them.  Also we can identify the "indexers" so we can allow
users to more easily select them.