Nicer EZ interface for selecting a channel list.
Sometimes one needs to associate a channel listing
with a recorder that cannot actually record from all the
channels in the list.  So the user can then pick a subset
from the full list.

This happens when a cable lineup has both analog and digital
channels.  Non-hybrid recorders can only do one or the
other.  So the same listing can be used but each recorder
needs to know the exact channels it can do or not do.

This new UI is a lot nicer has the available channels are
presented to the user to select.