Simplified the indexer services.
After having experience with many tries at transcoding and
simple wrapping of TS files, it was just easier to make one
indexer worker that was driven by a properties file.  This
removes alot of duplicated code and is much easier to add
more as it's simply an edit to a properties file.