The lightweight queue can become clogged.
A job that is designed to act upon a recording
file can be in a state of never executing if the
recording is deleted or hosed in some way.  We
try to execute the job upto a defined number of
times before giving up.  Lightweight jobs are
NOT critical so we will survive if they get
dropped.  But this state can stop all other
jobs from running.