Just hide video instances when missing.
Because people might have issues with mounting
drives or the need to migrate video files to other
paths we don't want to delete Video objects as it
would be nicer to just hide them.  When they restore
their drive or make changes the code will re-find
their old Video objects and un-hide them.

This is all dependent on video file names being unique
regardless of path.

This also has enabled us to do auto videoScans using the
autoart service.  This brings us closer to being able to
build this automatically allowing the user to just place
a file in the proper directory and we auto ingest them.