Fix bug with getting images and metadata for TV.
TV shows in the video library were being skipped because the
subcategory was not being set on initial scan.  Also more than
one episode was being skipped because they share the same artwork.
The fix for that was to make sure the ID was unique and added a FileID
that is used to write the URL - the URL episodes of the same show
share.