Make sure to check episodes that have no metadata.
Needed to add a check for TV episodes that might be new that
do not have ant description or metadata info.  If the user
adds new TV episodes they may not get populated because in
fanart new artwork is not needed.  So we also have to check
metadata for it to be flagged as needing checking.