Bug fix when user has deleted yet indexer works on.
Now that deletes are queued it can cause a timing problem.
The DB is updated but file are still on disk.  We were using
the disk file to determine if the recording still existed.

Now we have to get the old Recording from the DB to verify
it hasn't been deleted.