Scheduler should decide the Recordable Channels.
We moved some logic and code to the Scheduler to
determine the actual channels that are recordable
in this momemt of time.  The old code was not taking
into account sublists of Channels of Recorders.  So
the user could pick a Channel that isn't really
possible to Record.

In our setup we never saw this issue as the SD listing
was always the "truth".