Added the start of support for GTV.
We now have a servlet REST service that can be used to
get data and objects vis JSON serialization.  The interface
is not exactly the NMS but we will use it to model the
REST service API.

Also added a "state" object that will have useful information
in it so clients can determine the services available and
current state of all connected NMS servers.

Attempt at fixing path bug.  This bug happens because the
determination of a proper path for a recording had a snafu
because disk space could change between determining the path
and the start of the recording.  We now set this for "real"
just before commecing the recording.