Fix some list bugs.
When paging easpecially in Watch Recordings, the UI
is quite slow.  It's because a page is really a set
of move by one actions.  The UI was still updating
the screenshot and other info panels even though they
actually would not be drawn.  The lower level code
only fires these single moves if we are not in "page mode".
This greatly speeds things up.

The Schedule lists of RecordingRules and Upcomings had several
issues.  The old page selection text was being left "large"
and after editing a rule the state of the list was way off.
These are now fixed.