Update to latest timing framework.
I wanted to try to make the TV user interface a bit
more slick, fix some bugs and other things.  First
thing was to upgrade to the latest timing framework
since I hope it will be more efficient.

A lot of code changes were needed but now up to date.

A fix so the labels on the main screen no longer flicker.
Needed to set the alpha before animating.  Duh.

Still have some flicker with images on our 1080p TV.  Not
too bad though.  Nice on the 720p screens.

This commit really just fixes the above bug and syncs us
up with the timing framework from 2011 instead of 2008.