Try not to use JWindows over player windows.
There is a bug with painting the video and this is
eliminated by not using JWindows over the video.
Sometimes you get a "black bar" where the window
was when it went invisible.  This didn't happen with
mint 10 but happens often with mint 12.  This way of
adding to the player layeredpane will be better.