Live tv files need to be unique.
Now using Transfer to grab the live tv file means it must
be a unique file name.  It was only unique per channel and
LiveTV session but it causes Transfer issues.  Now it also
has a time stamp of when it was started - this will ensure
no name clashes.

Also fixed bug with hitting escape as the Transfer was cached
locally.  The user can hit escape and it will stop live tv when
the cache completes.  Not ideal because the user still has to wait
but at least it closes the session properly and all is cleaned
up properly.