Support for updating, restarting and player fixes.
We want it to be less work to update a running system. After
all a running system is just a set of OSGi bundles.  The actual
work is done via scripting and at this point we just have it
done for Linux.  Windows will be easy to do if rsync runs on Windows.
Have to check into that.

Also can restart the frontend via a menu pick.  Again this is
script based and only have the Linux version done.  This should be
rather simple for Windows though - just need a "sleep" command of
some sort.

Added a "maximize" function.  An attempt to be make fullscreen
video using chrome a little easier.  Not really happy with this since
each web site can invent their own way of doing it.