Added creating and maintaining favorite channels.
This is now possible with scheduling using the
"guide".  The user can toggle between all channels
and a favorite list.  The edit this list by selecting
a channel in the list.  If all channels are displayed,
then they can add the selected to a favorite list.  If
favorites are displayed then they can take them out.