Return to last live tv channel.
Best for the client to remember the last channel so
users can return to where they were instead of the
same start channel each time.

There could be a problem if that channel is not currently
available.  Just another good reason to avoid live TV - it's
just problematic.

But the current live tv code is pretty solid, at least here
at the labs.