Don't seek past the end of a recording.
This is simple to do for finished recordings but
it's annoying to seek to the end in case you are looking for
"next weeks coming up" and end the recording by a spurious
seek.  Then the bookmark is gone and a pain to reseek
to the end.

Still an issue with in progress recordings until the local
transfer copy has caught up to live.  So could still be a
bit annoying.