Add debugging and make ffmpeg mediainfo better.
Found out that running ffmpeg sometimes has funny
output.  Sometimes it has a bunch of more newlines
so parsing it by newlines is not a good idea.  Just
search for the video and audio tags and parse from
there.

While debugging realized I hadn't set the logging
service tracker so debug statement were not getting
to the log which made this harder to figure out.