Add missing items to install, make nice.
The idea of nice and ionice processing is Linux dependent
so we need to handle it properly.  On windows it just won't
be used.

Also added missing jar files to client/server install.  Testing
now with max memory put to 1024m.  Also make livetv dvr as the
default since much work has gone into it to improve things.