Ask user for scan type before doing it.
We found that the "auto" type for the hdhr doesn't really
work or it "sometimes" does the right thing.  So best for the
user to pick OTA or cable.  This also has the benefit of taking
a shorter amount of time since it will just do one or the other.