Make sure the guide for live tv is current.
Over time while watching live tv the guide info needs to be
updated or controlled so only show airings from "now" and in the
"future" are displayed.

This syncs up the scheduling as that code is called when a show is
in the future.

This probably still isn't perfect but it's the best we can get it
right now.