Fixed painting bug with video canvas.
What was happening on some machines was that the canvas
background was being over painted on the video canvas
after a resize.

It was odd as it was intermittant on different OSes
and version of X.

The solution was to override java.awt.Canvas with a new
Object and have the paint method do nothing.  After all,
all the pixels should be drawn by mplayer, not us.