Add some timing control to mplayer.
This is very hackish code but at this point not sure
how to fix it for real.  On the client-server install,
mplayer processes were not cleaning up correctly.  So I
put in a time out function before disposing of the mplayer
window which seems to fix it.  Since this is probably
related to a resource issue, a busier machine still might
fail - I've only tested it with live tv.  Who knows what
will happen when 4 tuners are working and the user is
watching recordings.

The default and fe installs this defaults to a zero wait
time so in essence the hack code is not executed.  It is
only executed on a > zero value in mplayer.properties.