Replace default properties with jflicks home.
In an effort to make a first install be more correct out
of the box we added a default storage to the install programs.
And on first run these settings should be set to the correct
path.  Probably not a long term setting but allows people
to try things without having to set much.