Handle 32-bit lirc lib better.
The installer now has 32-bit version for Linux that
installs the 32-bit lirc shared library.  Still need to
fix the build process as I just build a 32 bit lib in a
virtual box.  Have to figure out the magic gcc settings
to build it on my desktop machine.  It will do for now.