Remove comrat from the install.
Comrat works fine for us here but it does
complicate an install a little bit.  The TV UI
allows the user to pick both comrat and comskip (bug
as they should be radio buttons) and we also don't
have a nice way to update image files for comrat.  Nor
an easy way for users to add their own.

Just simpler to eliminate it for now.