Once Recordings sometimes invalid after restart.
Here was the problem:

Make a ONCE Recording.  When that Recording is ready to happen,
one of two states exist.  The guide data was refreshed or it
wasn't.  If it wasn't refreshed then the recording would happen.
If it was refreshed it would not.

The problem was that the ShowAiring cache was cleared out and on the
restart the Show and Airing was not there to set up the ONCE Recording.
Since the ONCE RecordingRule uses a ShowAiring instance to define it, we
noticed that it wasn't being persisted (perhaps).  We added CascadeOnUpdate
for the sched.dat database hoping this fixes it.  It may NOT as it needs
more testing.  But we saw that the RecordingRule has a ShowAiring
instance, it was the ShowAiring that did not have a Show or Airing
instance.

Hopefully this will fix it - the test will take two days to really
complete.