Added a static mplayer2 build.
Along with ffmpeg.exe, both programs are in the
windows installers.  The pain about the download
of both these builds, they are 7z files which by
default cannot be expanded on win7, or that is my
experience.  If they were zip files we would be fine
(winlirc is a zip for example) but we have enough
dependencies as it is then to include something
that can handle those files.