Making a cross-platform update service.
Instead of getting rsync working on windows, it's
better to have a Java way of updating our bundles.

I looked at the Felix OBR to do this in a "standard"
way but did not have much luck figuring it out and
having it work right.

So I made a home-grown method.  Up at the repository
we include a bundletree.xml file that defines the
dependencies we have defined.  At the current time this
file is hand-maintained - which was a lot of work to
build and something we will have to maintain.  A program
could be written to parse the bnd files and build it
dynamically which is something I should do.