Also look for bundle file size.
We want to update only new bundles and if they have
changed size.  It's possible that a new build of a
jar could be different but remain the same size in bytes,
but it's a bit remote.  And with this additional check,
we can do an ant clean, dodist and not cause updates to
happen when they really haven't.