Fix some bugs and add unknown channels.
Fixed some bugs with the GridGuide for live tv.

Also added code to the HDHomerun scan to include entries
for found channels but not configured by the user.  The
idea is that finding the channels right is kind of hard
and if we include these extras it makes editing the file
by hand a lot easier.  Usually just have to change the virtual
channel entry to what it is supposed to be.  And make sure
that channel is in your list via schedules direct.