The TV user interface effect changes.
We want to make the TV UI spifier but not at the cost of
looking ugly on lower CPU machines.  Unfortunately using
java it not the fastest (at least our code right now) to
do much fancy effects.  Like on our 1080p TV there is just
so much more pixels to push.  Doing image cross fades and the
like doesn't look great.

So we want to add more effects but then be able to turn them
off.

The only effect that is "on" all the time is the poster moving
animation.  It just "jumps" on slow machines anyway and doesn't
look horrible.  So we do it anyway.

Every thing else can be shutoff.