Make a UDP based mplayer instance.
To play a stream of UDP packets we have always used VLC because
it simply did the job.  MPlayer supports playing UDP but either
we couldn't figure it out or it needs some sort of protocol also
to be able to read them.

VLC does not perform well on the Atom processors.  Even with
vaapi installed, it just cannot hack playing this video or even
regular recordings well enough.  It does fine on fatser CPUs and
we will continue to use it.

So we want functionality like playing the Roku on these Atom
machine so we need to get mplayer to play these UDP streams.

We now receive the UDP packets in Java code and write them to
the stdin of a running mplayer.  We found with the right size
cache size that it works just about as good as VLC.  Because we
interact with the roku it's important not to have too large of
a cache size - button clicks will have a large latency (> 5 secs).
With a cache size of 2048 kb the delay is between 1-2 seconds.