Mplayer doing UDP, real recording length.
Now have mplayer playing udp by reading them
ourself and then sending over stdin for mplayer.
It works OK but a bit more delay than VLC. It
is usable though and allows for roku use on
atom processors.

Post recording job to accurately set the time
length of the recording.  We use ffmpeg to do
this now while we used to rely on timestamps of thread
starts and stops.  This is best now.

Also noticed that when many recordings end at the same
time, one would not "updateRecording" properly.  I
think this is a synchronize issue so I made that
method synchronized hoping this fixes the issue.  We
could have 6 threads trying to update at once.