Be able to have a preffered hour to get guide data.
User can now configure an hour of the day to run
the daily guide update.  This allows control when
this is done - like off peak hours.

Also added code to try to sense when a Recording
has failed and then to try to reschedule it.  This
failure has been happening on the HD-PVR where
we record commonly re-run episodes, often just
a few hours later.  Really shouldn't need this
because they should not be failing in the first place
but it's nice to have.