Be able to request new guide data.
Via the NMS be able to ask all ProgramData instances to
update to current guide data.  This is a way for a client
to update guide data to the most current.  We say "request"
because it can be refused if for example it is currently
running.

After it runs the ProgramData will run again at it's
preferred hour of the day.