The HDHRs firmware was updated (not my choice!).
The update broke the recording code.  One used to be able to select
the "program" by using the channel string.  Now that doesn't work.
So we have a new command line to get the current streams and we
figure out the proper integer program id which we can use to
select the right channel.

Overall this is better code but annoying that a firmware update took
away functionality.