First cut at LiveTV via REST.
This adds a service resource that can get the current channels
available via HLS and what it currently airing in this moment
in time.

Also added code to rebuild the Channel/ShowAiring hashmap
at the time a new schedule is downloaded.  This makes it
quick to request guide data later in the day.  It helps with
the new REST service so it is always quick.