Brought over older hdhomerun code that uses hdhome_config.
Looks like the java library was not setting the program info correctly
so the wrong program stream was being recorded.  I swear I tested
this before but another thing I stubbled upon that was broken.

This way it uses the old code only for the CONNECT model - I had
to have hdhomerun_config to save the data anyway so no biggie.